LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-valid-weekmonth-expected.txt


PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.valueMissing is true
